refactor: refactor api requests
